feat: rename engine to wrapper
