--- middleware/fiberEngine/hook.go
@@ -1,1 +1,0 @@
-package fiberEngine
