+++ middleware/fiberWrapper/hook.go
@@ -1,0 +1,1 @@
+package fiberWrapper
