+++ middleware/fiberWrapper/mux.go
@@ -1,4 +1,4 @@
-package fiberEngine
+package fiberWrapper
 
 import (
 	"errors"
